Start to specify how to serialize a ES value.
This starts building out the serialization infrastructure for sending
runtime-accessible objects over the wire. The format is based on
what's possible in CDP, with the notable change that the `value` field
is directly able to handle array and object types recursively rather
than using a seperate `preview` field. Also `type` and `subtype` are
merged into a single field and `value` and `unserializableValue` are
merged.

At this point there's no support for directly producing JSON values in
the serialization, we could add that later if required (but it's
arguably less important if the client is able to reconstruct the JSON
output when required).

There are also several other fields from CDP missing; we can add those
later if required.