Add a browsingContexts module
This contains a command to get all the descendent browsing contexts of
a given context, up to a specified depth. It also provides events to
monitor when contexts are created or destroyed.

This is the first module to contain any events, so some infrastructure
for emitting events is also added.