Make commands run in parallel. (#68)
Implement the decision to make all commands async so that there is no
guaranteed ordering between the responses of different commands, and
we don't have to block processing one command on the previous one
having sent the response.

To make this work I inlined the process a command steps so that only
actually running the command happens in parallel and not the initial
error checking. This also means that we don't need infrastructure to
communicate the parallel result back to the main event loop, but can
just write the response directly assuming response writing is
implicitly serialized.