Respond with an error for binary messages (#77)
Silently doing nothing is not very easy to debug, and more importantly
if in the future do support some binary messages, we should respond with
an error for any we do not recognize, just as we do for text messages.

Doing this required refactoring since "respond with an error" previous
assumed a text |data| argument, but we should not attempt to parse a
binary message as JSON.

Lift out the JSON parsing to before try to match against the CDDL, to
avoid parsing twice. Previously parsing was implicit in the "match"
verb, but we can just as well "match" against a parsed representation,
it is no more or less defined what that means.

One typo was fixed: occured

The only intended observable change is for non-text messages.