Allow emitting events on subscription
This adds support for backfilling events when subscribing to a new
event type either for the first time, or for a new set of
contexts. The use case is allowing local ends to update themselves with
the current state of the remote end, just by listening for events,
without needing to be very careful about the exact order for which
subscriptions happen.

If an event is backfilled, it must define remote end subscribe steps,
which take a set of browsing contexts where the event is newly active,
and an |include global| flag which indicates whether the subscription
is global.

These steps have a priority order so that we can create invariants
like browsing context events being emitted before the realms or log
messages relating to those contexts.

In order to make it easier to backfill events, this also changes the
subscription model to only allow subscibing per top-level context, not
per frame.