Add command and events related to browsing context navigation.
This adds a navigate command to the browsingContext module that
initiates a navigation of the given context to the given URL. The
command also accepts a wait parameter which affects whether it returns
immediately after the navigation is started or waits until the
document that's loaded is eventually parsed. In the case of early
return the navigation id allows later events to be connected to this
specific navigation.

In addition we also add a set of events relating to the navigation
lifecycle, corresponding to the navigation starting, failing, reaching
domContentLoaded, and reaching load. There is also an event for
synchronous same-page navigations.

There are quite a few open questions here about how we want to expose
error conditions. Currently we emit errors from the command, but these
don't have a lot of structured information and might end up with local
ends trying to parse additional information out of the message part.