Add a browsingContext.traverseHistory command
This traverses the history by a specified `delta` (e.g. +1 for forward,
or -1 for back).

For now there is no `wait` parameter; the command always returns as
soon as the hsitory traversal has been enqueued (this should
eventually be equivalent to setting wait to `none`).

This model differs somewhat from the CDP model where one navigates to
an explicit history entry and failure is only possible if the entry id
is invalid. But back/forwward only seems closer to the use cases we
have, and allowing events to be traced to a specific traversal seems
consistent with the way we handle other navigation-related events.