Add a task queue and use it for scheduling.
This adds a parallel task queue and uses it to schedule all events,
rather than using "in parallel" a lot. This makes it easier to reason
about the ordering of events.

In addition it makes all events happens as part of a task. This seems
likely to match implementations where events will presumably be
triggered from event listeners rather than synchronously as in the
spec at present. It also means that we don't need the event
suppression hack to ensure that navigation commands return before
related events are emitted.