Move exception handling into the script module.
This tightens up the handling of stack creation a little, although
it's still pretty loose because ECMAScript doesn't provide any help in
stack construction. We assume that given a completion record, we can
figure out a stack for the point at which that record was created, and
also a stack for the current point in the js execution.

It also makes a small normative change, introducing the StackTrace
type which has a callFrames property to better match CDP.