Add script.evaluate
This is like a minimal Runtime.evaluate from CDP; it allows running
script in an existing realm, with the return value being the value the
provided expression evaluates to.

If awaitPromise is not set to false, and the expression resolves as a
promise, we await promise resolution before returning the resolved value.

In the case that the script throws an exception (or we await a promise
that is rejected), the response contains the throw/reject value in the
result field and also contains an exceptionDetails field with details
of the exception.