Add basic API surface for sandboxed script execution.
The main change here is that the target type allows specifying a
sandbox name, and when a name is supplied, we got or create a sandbox
for the given context with the given name.

The details of how sandboxes actually work are intentionally elided
for now, since that's more complex to specify. But some non-normative
hints about their properties are given to sketch out the direction in
which things are moving.