Correct interpretation of null maximum depth
In the serialization steps for specific container types, when recursion
continues when "max depth" take the value "null". However, when
determining whether to begin serialization of nested data structures,
the value "null" was generally interpreted to mean "do not serialize."
For the case of Object values, null was not handled at all.

Update the algorithm to consistently interpret null as a signal for
unlimited recursion.