Add support for preload scripts (#292)
* Add support for "preload scripts"

These are scripts which are run when a realm (initially only for
realms with a Window global) is created. They allow a test harness to
inject functionality that's guaranteed to be available for any content
scripts that are subsequently loaded, and before any later scripts
that WebDriver injects into the context.

Registering a load script happens with the script.addPreloadScript
command. This returns a handle to the added script.

Load scripts can later be removed by passing the handle to
script.removePreloadScript.

Preload scripts can either be run directly in the newly created Window,
or in a named sandbox.

Co-authored-by: Henrik Skupin <mail@hskupin.info>