Move "Common data types" to the script module
The Common Data Types section was a bit of a relic from the early
organisation of the spec, before we clearly divided things into
modules and namespaced all the value types.

Since all the types defined in that section are related to the
representation of ECMAScript types in the protocol, it makes sense to
move the definitions into the script module, and so ensure that only
the top-level protocol definitions are unnamespaced.