Add network.BytesValue type
This provides a uniform representation of network data that may be
UTF-8 text or may be arbitary binary data. Non-UTF-8 data is always
encoded as base64 for transport.

This is a backwards-incompatible change since it coverts the
network.Cookie and network.Headers types from either having a `value`
or `binaryValue` field to always having a `value` field, but the value
being an object with a `type` field to distinguish the variants.