Fix CDDL
We need to use `/` instead of `//` as `//` are meant to union groups, not types. Also, we should be using groups when group discriminators are in use.

For example, `ResultData` is used as the type for `CommandResponse.result`, but `ResultData` is a group (since it uses `//`). Only types may be used as values.