Set the correct result type for browsingContext.Reload
The prose here has two different return statements, the first that
returns a `browsingContext.NavigateResult` and the second that returns
`null`.

Based on the history it seems like the intent was always to return
`browsingContext.NavigateResult`, and the `return null` was erronously
added during cleanup when it was noticed that the return type CDDL was
missing, and so was assuemd to be empty.

From a user perspective, having some way to track progress of a
reload, just like any other navigation, seems important, so this is
the right behaviour.