Support user prompt handlers for BiDi sessions
The key part of this is that with BiDi HTML calls into the spec when
it's going to show a prompt. So at that point we can see if there's a
handler defined, and if so return the handler to HTML.

All the actual logic to handle the dialogs ends up in the HTML spec.

From the BiDi point of view, the intended lifecycle is that you always
get the session.userPromptOpened and session.userPromptClosed events,
even if the prompt is automatically handled. But the
session.UserPromptOpened event gets a new `handler` property that
tells you whether the prompt will be automatically handled; if this is
set to "none" the automation has to send a session.handleUserPrompt
command (or wait for the prompt to be dismissed by the user).