Remove IteratorRemoteValue from the spec
Currently, how IteratorRemoteValue is not specified
as per https://github.com/w3c/webdriver-bidi/issues/537

Additionally, a general iterator is any object with a
next method as per https://tc39.es/ecma262/multipage/control-abstraction-objects.html#sec-iterator-interface

Since there is currently no use case requiring serialzation support
of iterators, we propose removing IteratorRemoteValue from the spec.

The following use cases have been considered:

1) In puppeteer code,

```
сonst it = await page.evaluateHandle(() => produce an iterator)
await it.evaluate(it => it.next())
```

Conclusion: serialization as an object is sufficient, sicne the user
knows if they are dealing with an iterator and additionally they can
check for the presence of the next method if needed.

2) Logging an iterator object to console

We believe that serializing an iterator as an object, would be
sufficient.