Replace HTML's "report the exception" with "report an exception". (#758)
The former is deprecated. Currently, the environment settings object's
realm's global is used (which is consistent with what the spec would
previously have said before). It's possible this ought to use the
sandbox's realm, but that realm is not currently used for the function
evaluation itself right now -- and that's the function whence the error
comes.

This is part of whatwg/html#10516.