Try to ensure sandboxed preload scripts run in the correct realm
Previously we were passing the environment settings object for the
Window realm into `evaluate function body`, which is incorrect.

Instead update `environment settings` to point to the environment
settings object for the sandboxed realm.

Also have a go at defining the steps to create a sandboxed realm and
ensure that it has an environment settings object that we can use.