Expose initiatorType and destination on network.Request
Currently we have network.Initiator, which exposes some detailed
information about the stack trace that generated the request on
network.beforeRequestSent, plus a `type` field that's not well
defined.

Fetch exposes `initiatorType` and `destination` and these are already
used by other standards e.g. performanceTiming. Having data that
directly corresponds to those other specs makes a good deal of sense
so that e.g. performance measurements using WebDriver have the same
fields with the same information as the DOM measurement APIs.

With this change we could remove `Network.Initiator.type` since that's
redundant with this data, but keep the field for extra information
about the stack that generated a request.