Fix broken fragment references in spec
This fixes broken and obsolete fragment references in the spec.

Most updates are straightforward. A few notes:
- `tools.ietf.org` now redirects to `datatracker.ietf.org`
- The update drops any link to "handled" in HTML. Not sure where the term was
defined but it no longer seems to exist.
- "default classic script fetch options" is "default script fetch options" in
practice. If "classic" matters, then that term no longer exists!

Not done as part of this update, but some of the references can likely be
handled directly by Bikeshed without having to define more brittle anchors.