Update Bikeshed metadata: add editors, TR link, drop level
This adds spec editors, as discussed in #817, adds the link to the published TR
draft, and drops the notion of level so that the spec gets published as
`webdriver-bidi` instead of `webdriver-bidi-1` (levels can be introduced later
on if that turns out to be needed).

PR to be merged after publication as First Public Working Draft, the TR link
won't exist before that.