Use `<code>` for references to a CDDL type (#831)
The spec wrapped the return type of commands in  `<pre class="cddl">`, even
when the return type is actually a reference to CDDL type defined elsewhere.

The convention with IDL blocks is that any `<pre class="idl">` block creates a
new IDL definition. I'd like to follow the same convention for CDDL.

On top of it, from a semantic perspective, such a reference is more "a fragment
of computer code" (`<code>`) than "a block of preformatted text" (`<pre>`).

Note: when and if Bikeshed starts supporting CDDL, this could be replaced by
something like `<a cddl-type>EmptyResult</a>` (or `{^EmptyResult^}` if the
shorthand gets adopted).