Adapt CDDL declarations to new version of Bikeshed (#927)
* Adapt CDDL declarations to new version of Bikeshed

Version 5.2.0 of Bikeshed released yesterday adds support for CDDL
highlighting and linking:
https://github.com/speced/bikeshed/issues/1773#issuecomment-2914475958
https://speced.github.io/bikeshed/#cddl

This update makes a minimal set of changes to adjust the CDDL declarations
to define the CDDL modules and blocks in a way that Bikeshed understands.
In particular, Bikeshed needs the blocks to reference the CDDL modules through
a `data-cddl-module` attribute.

Bikeshed now also generates a CDDL index at the end of the spec. This CDDL
index is divided into a remote end definition part and a local end definition
part. Each part contains all definitions defined as applying to the underlying
module. Generation of the index can be disabled if needed.

The definitions of the CDDL modules themselves require using a new
`cddl-module` definition type. Changing the definition type means that the
resulting IDs and the way to reference the definitions from other specs
changes. Specs that reference these definitions in WebDriver BiDi need a small
update accordingly.

* Adjust recently added CDDL block as well