feat: add userContext to protocol events
Since the spec added the ability to subscribe to events using user context IDs,
it becomes more difficult to link a given event handler with the user context
its event is initiated from. This change makes the event parameters more
ergonomic and able to make that link easier to process.

Fixes #1032.